Update class names
